requirements.txt
# Core Django
Django>=4.2.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1

# Database (SQLite for dev, prepare for PostgreSQL later)
psycopg2-binary>=2.9.7
mysqlclient>=2.2.0


# Authentication
djangorestframework-simplejwt>=5.3.0

# Blockchain Integration
web3>=6.11.0


# Environment Variables
python-dotenv==1.0.0

# Utilities
py-solc-x==1.1.1

django-filter==23.3  # For advanced filtering
drf-yasg==1.21.5     # For API documentation
gunicorn==21.2.0      # Production WSGI server
whitenoise==6.6.0     # For serving static files in production
qrcode[pil]==7.4.2
cryptography==41.0.7
drf-spectacular==0.27.1